Fix command annotation value keys and slider step type

+++ src/main/kotlin/club/sk1er/mods/levelhead/config/LevelheadConfig.kt
@@ -37,7 +37,7 @@ object LevelheadConfig : Config(Mod("BedWars Levelhead", ModType.HYPIXEL), "bedw
     @Text(name = "Proxy Auth Token", secure = true)
     var proxyAuthToken: String = ""
 
-    @Slider(name = "Star Cache TTL (minutes)", min = MIN_STAR_CACHE_TTL_MINUTES.toFloat(), max = MAX_STAR_CACHE_TTL_MINUTES.toFloat(), step = 1f)
+    @Slider(name = "Star Cache TTL (minutes)", min = MIN_STAR_CACHE_TTL_MINUTES, max = MAX_STAR_CACHE_TTL_MINUTES, step = 1)
     var starCacheTtlMinutes: Int = DEFAULT_STAR_CACHE_TTL_MINUTES
 
     @Switch(name = "Use Threat Colors", description = "Color stats based on FKDR")
